requirements.txt
# Core dependencies for Yad2 Haifa Property Monitor
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
